updates the entire specification and all the docs!
